+++ data-raw/create-package-data.R
@@ -2,11 +2,13 @@
 
 cat(getwd(), "\n")
 
-source("data-raw/create-COSMICv3-SBS96-signatures.R")
+source("data-raw/create-COSMICv3-signatures.R")
 source("data-raw/create-COSMICv3-sigs-URL.R")
 source("data-raw/create-COSMICv3-sigs-aetiology-info.R")
 
 usethis::use_data(COSMIC.v3.genome.SBS96.sigs,
+                  COSMIC.v3.hg19.genome.ID.sigs,
+                  COSMIC.v3.sigs,
                   COSMIC.v3.SBS96.sig.links, 
                   COSMIC.v3.SBS192.sig.links, 
                   COSMIC.v3.DBS78.sig.links,